move yaml parsing/dumping to base model
