+++ great_expectations/experimental/datasources/config.py
@@ -1,16 +1,17 @@
 """POC for loading config."""
 from __future__ import annotations
 
-import json
 import logging
 import pathlib
-from io import StringIO
 from pprint import pformat as pf
-from typing import Dict, Type, Union, overload
+from typing import Dict, Type, Union
 
-from pydantic import BaseModel, validator
+from pydantic import validator
 from ruamel.yaml import YAML
 
+from great_expectations.experimental.datasources.base_config_models import (
+    ExperimentalBaseModel,
+)
 from great_expectations.experimental.datasources.interfaces import Datasource
 from great_expectations.experimental.datasources.sources import _SourceFactories
 
@@ -22,10 +23,8 @@ yaml.default_flow_style = False
 
 LOGGER = logging.getLogger(__name__)
 
-# TODO (kilo59): should we move yaml parsing/dumping to ExperimentalBase??
 
-
-class GxConfig(BaseModel):
+class GxConfig(ExperimentalBaseModel):
     datasources: Dict[str, Datasource]
 
     @classmethod
@@ -57,32 +56,3 @@ class GxConfig(BaseModel):
 
         LOGGER.info(f"Loaded 'datasources' ->\n{repr(loaded_datasources)}")
         return loaded_datasources
-
-    @overload
-    def yaml(self, stream_or_path: Union[StringIO, None] = None, **yaml_kwargs) -> str:
-        ...
-
-    @overload
-    def yaml(self, stream_or_path: pathlib.Path, **yaml_kwargs) -> pathlib.Path:
-        ...
-
-    def yaml(
-        self, stream_or_path: Union[StringIO, pathlib.Path, None] = None, **yaml_kwargs
-    ) -> Union[str, pathlib.Path]:
-        """
-        Serialize the config object as yaml.
-
-        Writes to a file if a `pathlib.Path` is provided.
-        Else it writes to a stream and returns a yaml string.
-        """
-        if stream_or_path is None:
-            stream_or_path = StringIO()
-
-        # pydantic json encoder has support for many more types
-        # TODO: can we dump json string directly to yaml.dump?
-        intermediate_json = json.loads(self.json())
-        yaml.dump(intermediate_json, stream=stream_or_path, **yaml_kwargs)
-
-        if isinstance(stream_or_path, pathlib.Path):
-            return stream_or_path
-        return stream_or_path.getvalue()
